#000: Vinay : WIP location decoders
