+++ lib/test/clients/location/LocationPostService.unit.test.ts
@@ -1,10 +1,11 @@
 import { mocked } from 'ts-jest/utils'
 import { AkkaLocation, Duration, HttpConnection } from '../../../src/clients/location'
 import { LocationServiceImpl } from '../../../src/clients/location/LocationServiceImpl'
-import { DoneD, LocationListD } from '../../../src/clients/location/models/LocationResponses'
+import { DoneD } from '../../../src/clients/location/models/LocationResponses'
 import * as Req from '../../../src/clients/location/models/PostCommand'
 import type { LocationHttpMessage } from '../../../src/clients/location/models/PostCommand'
 import type { LocationWebSocketMessage } from '../../../src/clients/location/models/WsCommand'
+import { LocationListD } from '../../../src/decoders/LocationDecoders'
 import { Prefix } from '../../../src/models'
 import { HttpTransport } from '../../../src/utils/HttpTransport'
 import { Ws } from '../../../src/utils/Ws'
