+++ lib/test/clients/location/LocationWsService.unit.test.ts
@@ -2,8 +2,8 @@ import { mocked } from 'ts-jest/utils'
 import { HttpConnection } from '../../../src/clients/location'
 import { LocationServiceImpl } from '../../../src/clients/location/LocationServiceImpl'
 import type { LocationHttpMessage } from '../../../src/clients/location/models/PostCommand'
-import { TrackingEventD } from '../../../src/clients/location/models/TrackingEvent'
 import { LocationWebSocketMessage, Track } from '../../../src/clients/location/models/WsCommand'
+import { TrackingEventD } from '../../../src/decoders/LocationDecoders'
 import { Prefix } from '../../../src/models'
 import { HttpTransport } from '../../../src/utils/HttpTransport'
 import { Ws } from '../../../src/utils/Ws'
